feat: add TwoSimplePieChart example in TypeScript
